+++ fairroot/parbase/FairParSet.cxx
@@ -31,11 +31,12 @@ FairParSet::FairParSet(const char* name, const char* title, const char* context,
     , status(kFALSE)
     , changed(kFALSE)
     , owned(owner)
+    , creationMode(false)
     , paramContext(context)
     , author("")
     , description("")
 {
-    for (Int_t i = 0; i < 3; i++) {
+    for (int i = 0; i < 3; i++) {
         versions[i] = -1;
     }
 }
@@ -50,7 +51,7 @@ Bool_t FairParSet::init()
     // FairRunAna* fRun =FairRunAna::Instance();
     // cout << "-I-  FairParSet::init() " << GetName() << endl;
 
-    Bool_t allFound = kFALSE;
+    bool allFound = false;
     FairParIo* io = 0;
     if (rtdb) {
         io = rtdb->getFirstInput();
@@ -68,6 +69,10 @@ Bool_t FairParSet::init()
         }
     }
     if (allFound) {
+        return kTRUE;
+    }
+    if (creationMode) {
+        LOG(debug) << "init() " << GetName() << "[creationMode=ON] not initialized.";
         return kTRUE;
     }
     LOG(error) << "init() " << GetName() << " not initialized";
@@ -91,33 +96,25 @@ Int_t FairParSet::write()
 void FairParSet::print()
 {
     // prints information about container (versions,status,hasChanged...)
-    cout << "-----  " << GetName() << "  -----" << '\n';
+    LOG(info) << "-----  " << GetName() << "  -----";
     if (!paramContext.IsNull()) {
-        cout << "Context/Purpose:       " << paramContext << '\n';
+        LOG(info) << "Context/Purpose:       " << paramContext;
     }
     if (!author.IsNull()) {
-        cout << "Author:                " << author << '\n';
+        LOG(info) << "Author:                " << author;
     }
     if (!description.IsNull()) {
-        cout << "Description:           " << description << '\n';
+        LOG(info) << "Description:           " << description;
     }
-    cout << "first input version:   " << versions[1] << '\n';
-    cout << "second input version:  " << versions[2] << '\n';
-    if (changed) {
-        cout << "has changed" << '\n';
-    } else {
-        cout << "has not changed" << '\n';
-    }
-    if (status) {
-        cout << "is static" << '\n';
-    } else {
-        cout << "is not static" << '\n';
-    }
+    LOG(info) << "first input version:   " << versions[1];
+    LOG(info) << "second input version:  " << versions[2];
+    LOG(info) << "has" << (changed ? "" : " not") << " changed";
+    LOG(info) << "is" << (status ? "" : " not") << " static";
 }
 
 void FairParSet::clear()
 {
-    status = kFALSE;
+    status = false;
     resetInputVersions();
 }
 
@@ -125,10 +122,10 @@ void FairParSet::resetInputVersions()
 {
     // resets the input versions if the container is not static
     if (!status) {
-        for (Int_t i = 0; i < 3; i++) {
+        for (int i = 0; i < 3; i++) {
             versions[i] = -1;
         }
-        changed = kFALSE;
+        changed = false;
     }
 }
 
@@ -150,7 +147,7 @@ FairParSet::FairParSet(const FairParSet& from)
  fTitle   = from.fTitle;
  detName  = from.detName;
 */
-    for (Int_t i = 0; i < 3; i++)
+    for (int i = 0; i < 3; i++)
         versions[i] = from.versions[i];
     /*
  status   = from.status;