+++ fairroot/parbase/FairParSet.h
@@ -20,16 +20,17 @@ class FairParSet : public TObject
     TString fName;          //
     TString fTitle;         //
     TString detName;        //! name of the detector the container belongs to
-    Int_t versions[3];      //! versions of container in the 2 possible inputs
-    Bool_t status;          //! static flag
-    Bool_t changed;         //! flag is kTRUE if parameters have changed
-    Bool_t owned;           //! if flag is KTRUE FairDB has the par. class ownership
+    int versions[3];        //! versions of container in the 2 possible inputs
+    bool status;            //! static flag
+    bool changed;           //! flag is kTRUE if parameters have changed
+    bool owned;             //! if flag is KTRUE FairDB has the par. class ownership
+    bool creationMode;      //! allow creation mode, it suppresses error message in the init
     TString paramContext;   // Context/purpose for parameters and conditions
     TString author;         // Author of parameters
     TString description;    // Description of parameters
 
   public:
-    FairParSet(const char* name = "", const char* title = "", const char* context = "", Bool_t owner = kFALSE);
+    FairParSet(const char* name = "", const char* title = "", const char* context = "", bool owner = false);
     virtual ~FairParSet() {}
 
     const char* GetName() const override { return fName.Data(); }
@@ -45,13 +46,13 @@ class FairParSet : public TObject
     const char* getDetectorName() { return detName.Data(); }
     void resetInputVersions();
 
-    void setInputVersion(Int_t v = -1, Int_t i = 0)
+    void setInputVersion(int v = -1, int i = 0)
     {
         if (i >= 0 && i < 3) {
             versions[i] = v;
         }
     }
-    Int_t getInputVersion(Int_t i)
+    int getInputVersion(int i)
     {
         if (i >= 0 && i < 3) {
             return versions[i];
@@ -60,14 +61,17 @@ class FairParSet : public TObject
         }
     }
 
-    void setStatic(Bool_t flag = kTRUE) { status = flag; }
-    Bool_t isStatic() { return status; }
+    void setStatic(bool flag = kTRUE) { status = flag; }
+    bool isStatic() { return status; }
 
-    void setOwnership(Bool_t flag = kTRUE) { owned = flag; }
-    Bool_t isOwned() { return owned; }
+    void setOwnership(bool flag = kTRUE) { owned = flag; }
+    bool isOwned() { return owned; }
 
-    void setChanged(Bool_t flag = kTRUE) { changed = flag; }
-    Bool_t hasChanged() { return changed; }
+    void setChanged(bool flag = kTRUE) { changed = flag; }
+    bool hasChanged() { return changed; }
+
+    void setCreationMode(bool flag = true) { creationMode = flag; }
+    bool isCreationMode() { return creationMode; }
 
     const char* getParamContext() const { return paramContext.Data(); }
 
@@ -83,16 +87,15 @@ class FairParSet : public TObject
         description = r.getDescription();
     }
 
-    // TODO These two methods are not used in FairRoot at all.
-    // They probably should be marked deprecated (or final, or = delete)
-    // and later removed.
-    virtual void fill(UInt_t){};
-    virtual void store(UInt_t){};
+    // These two methods are not used in FairRoot at all.
+    // Therefore mark them deprecated.
+    [[deprecated("fill function is not used, report back to FairRoot team should you need the function")]] virtual void fill(UInt_t){};
+    [[deprecated("store function is not used, report back to FairRoot team should you need the function")]] virtual void store(UInt_t){};
 
     FairParSet& operator=(const FairParSet&);
     FairParSet(const FairParSet&);
 
-    ClassDefOverride(FairParSet, 2);   // Base class for all parameter containers
+    ClassDefOverride(FairParSet, 3);   // Base class for all parameter containers
 };
 
 #endif /* !FAIRPARSET_H */